Reconnect on more conditions
Previously we requires OSError with errno as a condition for reconnect.
But this commit allows toreconnect on manually thrown errors such as
`ConnectionAbortedError`, and also ones thrown protocol `connect()`
method (not just on socket connect method).
